# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SafeAccountCurrentState = typeof SafeAccountCurrentState[keyof typeof SafeAccountCurrentState];


export const SafeAccountCurrentState = {
  INIT: 'INIT',
  AUTH_PENDING: 'AUTH_PENDING',
  AUTHENTICATED: 'AUTHENTICATED',
  NEGOTIATION: 'NEGOTIATION',
  ACTION: 'ACTION',
  ESCALATED: 'ESCALATED',
  CALL_ENDED: 'CALL_ENDED',
} as const;

export interface SafeAccount {
  accountId: string;
  customerName: string;
  /** @nullable */
  loanType: string | null;
  /** @nullable */
  overdueAmount: number | null;
  /** @nullable */
  daysPastDue: number | null;
  authenticated: boolean;
  currentState: SafeAccountCurrentState;
}

export interface ToolCallRecord {
  id: string;
  toolName: string;
  state: string;
  success: boolean;
  createdAt: string;
  summary: string;
  latencyMs: number;
}

export type DispositionStatus = typeof DispositionStatus[keyof typeof DispositionStatus];


export const DispositionStatus = {
  PTP_AGREED: 'PTP_AGREED',
  ALREADY_PAID: 'ALREADY_PAID',
  DISPUTED: 'DISPUTED',
  HARDSHIP_ESCALATED: 'HARDSHIP_ESCALATED',
  WRONG_PERSON: 'WRONG_PERSON',
  DO_NOT_CALL: 'DO_NOT_CALL',
  CALLBACK_REQUEST: 'CALLBACK_REQUEST',
  VERIFICATION_FAILED: 'VERIFICATION_FAILED',
  NO_RESPONSE: 'NO_RESPONSE',
  HOSTILE_CALLER: 'HOSTILE_CALLER',
} as const;

export interface Disposition {
  id: string;
  accountId: string;
  status: DispositionStatus;
  notes: string;
  createdAt: string;
}

export interface PtpRecord {
  id: string;
  accountId: string;
  date: string;
  amount: number;
  paymentLinkSent: boolean;
  createdAt: string;
}

export interface DashboardMetrics {
  totalCalls: number;
  successfulPtp: number;
  ptpRate: number;
  escalations: number;
  escalationRate: number;
  dncCount: number;
  alreadyPaidCount: number;
  disputeCount: number;
  hardshipCount: number;
  noResponseCount: number;
  verificationFailureRate: number;
  containmentRate: number;
  dispositionCompletionRate: number;
  averageToolLatencyMs: number;
  averageCallDurationMs: number;
  toolSuccessRate: number;
}

export interface VoiceConfiguration {
  name: string;
  description: string;
  provider: string;
  /** @nullable */
  voiceId: string | null;
  status: string;
  referenceFile: string;
}

export type DashboardSnapshotCurrentState = typeof DashboardSnapshotCurrentState[keyof typeof DashboardSnapshotCurrentState];


export const DashboardSnapshotCurrentState = {
  INIT: 'INIT',
  AUTH_PENDING: 'AUTH_PENDING',
  AUTHENTICATED: 'AUTHENTICATED',
  NEGOTIATION: 'NEGOTIATION',
  ACTION: 'ACTION',
  ESCALATED: 'ESCALATED',
  CALL_ENDED: 'CALL_ENDED',
} as const;

export type DashboardSnapshotAuthenticationStatus = typeof DashboardSnapshotAuthenticationStatus[keyof typeof DashboardSnapshotAuthenticationStatus];


export const DashboardSnapshotAuthenticationStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  FAILED: 'FAILED',
} as const;

export interface DashboardSnapshot {
  sessionId: string;
  agentName: string;
  companyName: string;
  firstMessage: string;
  account: SafeAccount;
  currentState: DashboardSnapshotCurrentState;
  authenticationStatus: DashboardSnapshotAuthenticationStatus;
  authAttempts: number;
  toolCalls: ToolCallRecord[];
  dispositions: Disposition[];
  ptpRecords: PtpRecord[];
  metrics: DashboardMetrics;
  voice: VoiceConfiguration;
}

export type DemoActionInputAction = typeof DemoActionInputAction[keyof typeof DemoActionInputAction];


export const DemoActionInputAction = {
  VERIFY: 'VERIFY',
  SUCCESSFUL_PTP: 'SUCCESSFUL_PTP',
  ALREADY_PAID: 'ALREADY_PAID',
  DISPUTE: 'DISPUTE',
  DNC: 'DNC',
  HARDSHIP: 'HARDSHIP',
  WRONG_PERSON: 'WRONG_PERSON',
  HOSTILE_CALLER: 'HOSTILE_CALLER',
  NO_RESPONSE: 'NO_RESPONSE',
} as const;

export interface DemoActionInput {
  action: DemoActionInputAction;
}

export type WebhookToolCallFunctionArguments = { [key: string]: unknown };

export type WebhookToolCallFunction = {
  name: string;
  arguments: WebhookToolCallFunctionArguments;
};

export interface WebhookToolCall {
  id: string;
  function: WebhookToolCallFunction;
}

export type WebhookInputMessageType = typeof WebhookInputMessageType[keyof typeof WebhookInputMessageType];


export const WebhookInputMessageType = {
  'tool-calls': 'tool-calls',
} as const;

export type WebhookInputMessageCall = {
  id: string;
};

export type WebhookInputMessage = {
  type: WebhookInputMessageType;
  call?: WebhookInputMessageCall;
  toolCalls: WebhookToolCall[];
};

export interface WebhookInput {
  message: WebhookInputMessage;
}

export interface WebhookResult {
  toolCallId: string;
  result: string;
}

export interface WebhookResponse {
  results: WebhookResult[];
}

export interface ErrorResponse {
  error: string;
  code?: string;
}